가변성을 Semi_Discrete_Equation 수준에서 다루도록 수정

+++ MS_Solver/SRC/Matrix.cpp
@@ -15,7 +15,7 @@ Matrix Matrix_Base::operator*(const double constant) const
 	const auto num_values = this->num_values();
 	std::vector<double> values(num_values);
 
-	const auto n = num_values;
+	const auto n = static_cast<int>(num_values);
 	const auto incx = 1;
 	cblas_dscal(n, constant, values.data(), incx);
